+++ hotel/views/hotel_room_amenities.xml
@@ -21,7 +21,7 @@
 									<field name="type" select="2" string="Amenity Type" />
 									<field name="product_manager" select="2" string="Amenity manager" />
 									<field name="state" select="2" />
-									<field name="categ_id" select="1" domain="[('istype','=',True)]" />
+									<field name="categ_id" select="1" domain="[('isamenitype','=',True)]" />
 									<field name="uom_id" />
 									<field name="uom_po_id" />
 								</group>
